style(evals): drop a redundant int() cast, use an explicit !s conversion flag
Two more cosmetic findings from the ruff sweep, same style as the
earlier measure_drift_trace.py fix:

- perf.py's _percentile(): round(x) with one argument already returns
  int in Python 3, so wrapping it in int(...) was a no-op. Verified
  _percentile() unchanged (median/p95/empty-list cases all match).
- sweep.py: str(p.value) inside an f-string -> {p.value!s}, identical
  output.

No behavior change in either.

+++ evals/perf.py
@@ -81,7 +81,7 @@ def _percentile(values: list[float], q: float) -> float:
     if not values:
         return 0.0
     s = sorted(values)
-    k = min(len(s) - 1, int(round(q * (len(s) - 1))))
+    k = min(len(s) - 1, round(q * (len(s) - 1)))
     return s[k]
